+++ include/chemfiles/utils.hpp
@@ -19,7 +19,9 @@ inline std::vector<std::string> split(const std::string &s, char delim) {
     std::string item;
     std::vector<std::string> elems;
     while (std::getline(ss, item, delim)) {
-        elems.push_back(item);
+        if (item != "") {
+            elems.push_back(item);
+        }
     }
     return elems;
 }