+++ src/FormatFactory.cpp
@@ -7,6 +7,7 @@
 #include "chemfiles/formats/Molfile.hpp"
 #include "chemfiles/formats/AmberNetCDF.hpp"
 #include "chemfiles/formats/LAMMPSData.hpp"
+#include "chemfiles/formats/Tinker.hpp"
 #include "chemfiles/formats/PDB.hpp"
 #include "chemfiles/formats/XYZ.hpp"
 #include "chemfiles/formats/TNG.hpp"
@@ -36,6 +37,9 @@ FormatFactory::FormatFactory() : formats_(), extensions_() {
 
     this->register_name<AmberNetCDFFormat>("Amber NetCDF");
     this->register_extension<AmberNetCDFFormat>(".nc");
+
+    this->register_name<TinkerFormat>("Tinker");
+    this->register_extension<TinkerFormat>(".arc");
 
     this->register_name<LAMMPSDataFormat>("LAMMPS Data");