+++ CHANGELOG.md
@@ -30,8 +30,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Amber NetCDF format is now activated by default, by embedding the netcdf
   library in chemfiles.
 * Added `LAMMPS Data` format, reading and writing [LAMMPS data files].
-* Added mol2 file format support using VMD molfiles plugin;
-* Added Molden file format support using VMD molfiles plugin;
+* Added `Tinker` format, reading and writing Tinker XYZ file format.
+* Added `MOL2` format, reading mol2 files using VMD molfiles plugin.
+* Added `Molden` format, reading molden files using VMD molfiles plugin.
 
 [LAMMPS data files]: http://lammps.sandia.gov/doc/read_data.html